requirements.txt
astroid==1.6.4
click==6.7
cycler==0.10.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
numpy==1.14.4
Pillow==5.1.0
pylint==1.9.1
PyMySQL==0.8.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
wordcloud==1.4.1
wrapt==1.10.11
